move shared functions to data service for optimized delivery
