+++ OKP-2-frontend/src/app/components/list-events/list-events.component.html
@@ -15,4 +15,9 @@
       </div>
     </div>
   </section>
+
+  <div *ngIf="!events || events.length === 0">
+    No events found.
+  </div>
+
 </div>
