+++ OKP-2-frontend/src/app/components/list-places/list-places.component.html
@@ -15,4 +15,10 @@
       </div>
     </div>
   </section>
+
+  <div *ngIf="!places || places.length === 0">
+    No places found.
+  </div>
+
 </div>
+
