+++ OKP-2-frontend/src/app/components/map/map.component.ts
@@ -109,7 +109,15 @@ export class MapComponent implements OnInit {
   }
 
   constructor(private dataService: DataService,
-              private sharedService: SharedService) {}
+              private sharedService: SharedService
+              ) {
+                this.activities = dataService.activities;
+                this.activityLocations = dataService.activityLocations;
+                this.events = dataService.events;
+                this.eventLocations = dataService.eventLocations;
+                this.places = dataService.places;
+                this.placeLocations = dataService.placeLocations;
+              }
 
   ngOnInit(): void {
     // get user's current location and keep track of it via geolocation api
@@ -131,9 +139,9 @@ export class MapComponent implements OnInit {
         this.updateLocation();
         console.log(this.center);
 
-        this.getActivitiesData();
-        this.getEventsData();
-        this.getPlacesData();
+        this.dataService.getActivitiesData();
+        this.dataService.getEventsData();
+        this.dataService.getPlacesData();
       },
       (error) => {
         console.log('Error getting user location:', error.message);
@@ -142,102 +150,14 @@ export class MapComponent implements OnInit {
         this.updateLocation();
         console.log(this.center);
 
-        this.getActivitiesData();
-        this.getEventsData();
-        this.getPlacesData();
+        this.dataService.getActivitiesData();
+        this.dataService.getEventsData();
+        this.dataService.getPlacesData();
 
       });
     } else {
       console.log('Geolocation is not supported by this browser.');
     }
-  }
-
-  getActivitiesData() {
-    this.dataService.getActivities().subscribe((res: APIResponse) => {
-        let result = JSON.parse(res.data.result);
-        this.activities = result.rows;
-        console.log(this.activities);
-
-        // fetch the location coordinates and push them to an array
-        for (const activity of this.activities) {
-            const { lat, long } = activity.address?.location ?? {};
-            const name = activity.descriptions["fi"]?.name ?? activity.descriptions["en"]?.name;
-            const { streetName, postalCode, city } = activity.address;
-            const image = activity.media[0]?.originalUrl;
-            const url = activity.storeUrl;
-            const about = activity.descriptions["fi"]?.description ?? activity.descriptions["en"]?.description;
-            const tags = activity.tags;
-            this.activityLocations.push({
-              position: { lat, lng: long },
-              name,
-              address: { street_address: streetName, postal_code: postalCode, city},
-              image,
-              url,
-              about,
-              tags
-            });
-            }
-            console.log(this.activityLocations);
-    });
-  }
-
-  getEventsData() {
-    this.dataService.getEvents().subscribe((res: APIResponse) => {
-        let result = JSON.parse(res.data.result);
-        this.events = result.data;
-        console.log(this.events);
-
-        // fetch the location coordinates and push them to an array
-        for (const event of this.events) {
-        const { lat, lon } = event.location ?? {};
-        const name = event.name.fi ?? event.name.en;
-        const { street_address, postal_code, locality } = event.location.address;
-        const image = event.description.images.length > 0 ? event.description.images[0].url : null;
-        const url = event.info_url;
-        const about = event.description.intro;
-        const tags = event.tags.map(tag => tag.name);
-        this.eventLocations.push({
-          position: { lat, lng: lon },
-          name,
-          address: { street_address, postal_code, city: locality},
-          image,
-          url,
-          about,
-          tags
-        });
-        }
-        console.log(this.eventLocations);
-    });
-  }
-
-  getPlacesData() {
-    this.dataService.getPlaces().subscribe((res: APIResponse) => {
-        let result = JSON.parse(res.data.result);
-        this.places = result.data;
-        console.log(this.places);
-
-        // fetch the location coordinates and push them to an array
-        for (const place of this.places) {
-            const { lat, lon } = place.location ?? {};
-            const name = place.name.fi ?? place.name.en;
-            const { street_address, postal_code, locality } = place.location.address;
-            // const image = place.description.images[0].url;
-            const image = place.description.images.length > 0 ? place.description.images[0].url : null;
-            const url = place.info_url;
-            const about = place.description.intro;
-            const tags = place.tags.map(tag => tag.name);
-            this.placeLocations.push({
-              position: { lat, lng: lon },
-              name,
-              address: { street_address, postal_code, city: locality},
-              image,
-              url,
-              about,
-              tags
-            });
-            }
-            console.log(this.placeLocations);
-    });
   }
 
   openInfoWindow(marker, activity) {
